RxRPC control messages working (hopefully ;)).

+++ kernel/castle_public.h
@@ -2,7 +2,7 @@
 #define __CASTLE_PUBLIC_H__
 
 typedef uint32_t slave_uuid_t;
-typedef uint64_t snap_id_t;
+typedef uint32_t snap_id_t;
 typedef uint32_t region_id_t;
 typedef uint32_t transfer_id_t;
 
@@ -118,26 +118,26 @@ typedef struct castle_control_cmd_region_create {
     slave_uuid_t slave;        /* IN  */
     snap_id_t    snapshot;     /* IN  */
     uint32_t     start;        /* IN  */
-    uint32_t length;           /* IN  */
-    int ret;                   /* OUT */
-    region_id_t id;            /* OUT */
+    uint32_t     length;       /* IN  */
+    int          ret;          /* OUT */
+    region_id_t  id;           /* OUT */
 } cctrl_cmd_region_create_t;
 
 typedef struct castle_control_cmd_region_destroy {
     region_id_t id;            /* IN */
-    int ret;                   /* OUT */
+    int         ret;           /* OUT */
 } cctrl_cmd_region_destroy_t;
 
 typedef struct castle_control_cmd_transfer_create {
-    snap_id_t snapshot;        /* IN  */
-    int direction;             /* IN  */
-    int ret;                   /* OUT */
+    snap_id_t     snapshot;    /* IN  */
+    uint32_t      direction;   /* IN  */
+    int           ret;         /* OUT */
     transfer_id_t id;          /* OUT */
 } cctrl_cmd_transfer_create_t;
 
 typedef struct castle_control_cmd_transfer_destroy {
-    transfer_id_t id;            /* IN */
-    int ret;                   /* OUT */
+    transfer_id_t id;          /* IN  */
+    int           ret;         /* OUT */
 } cctrl_cmd_transfer_destroy_t;
 
 typedef struct castle_control_ioctl {
